Added texture to bumpMap property
Each Mesh has a Material and the Material has a property called bumpMap -- this affects how light and shadows are projected on the object. Adding our texture to the bumpMap property of the material of the meshes makes gives the meshes more dimension, which in this case adds to the "3D pixel art" quality that we're going for. Cool!